% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ods_sheets}
\alias{ods_sheets}
\title{List the sheets in an ODS file}
\usage{
ods_sheets(path)
}
\arguments{
\item{path}{The ODS file}
}
\value{
A character vector of sheet names
}
\description{
Get a list of sheets in the ODS file, either to explore the structure of a
file or to use as an input for iterating over a whole spreadsheet document.
}
\examples{
example <- system.file("extdata", "basic_example.ods", package = "tidyods")
ods_sheets(example)
}
