% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectify.R
\name{simple_rectify}
\alias{simple_rectify}
\alias{smart_rectify}
\title{Convert ODS cells to a two-dimensional dataset}
\usage{
simple_rectify(
  ods_cells,
  skip = 0,
  col_headers = TRUE,
  values_from = "base_value"
)

smart_rectify(ods_cells)
}
\arguments{
\item{ods_cells}{A set of cells from \code{read_ods_cells()}.}

\item{skip}{The number of rows to skip before extracting the sheet.}

\item{col_headers}{Whether to use the first row (after any skipping) as the
column header (\code{TRUE} is the default), alternatively a character vector of
column names can be provided.}

\item{value_from}{The column from \code{ods_cells} to output values from.}
}
\value{
A tibble representing the original spreadsheet format.
}
\description{
There are two functions to "rectify" a set of ODS cells extracted by
\code{\link[=read_ods_cells]{read_ods_cells()}} back into a traditional two-dimensional spreadsheet
rectangle.
}
\details{
\code{simple_rectify()} will perform a basic reshaping of the dataset, by default
it will use the \code{base_value} extracted by \code{read_ods_cells()} but this can
be changed to a different column by setting the \code{values_from} argument. The
rectifier will also by default try to use the first row for column headers,
alternatively you can provide your own column names or set
\code{col_headers = FALSE} to get generic column names of \code{x1}, \code{x2}, etc.

\code{smart_rectify()} performs a more complex reshaping of the dataset, by
guessing the location of column headers and using the \code{value_type}
information generated by \code{\link[=read_ods_cells]{read_ods_cells()}} to determine whether columns can
be coerced to a non-string data type (either numeric, logical, date or time).
}
\examples{
example <- system.file("extdata", "basic_example.ods", package = "tidyods")
example_cells <- read_ods_cells(example, 1)
simple_rectify(example_cells, base_values = TRUE)
smart_rectify(example_cells)
}
