% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectify.R
\name{simple_rectify}
\alias{simple_rectify}
\title{Convert ODS cells to a two-dimensional dataset}
\usage{
simple_rectify(ods_cells, base_values = TRUE)
}
\arguments{
\item{ods_cells}{A set of cells from an ODS sheet}

\item{base_values}{Whether to use the base_value of a cell (TRUE, the default)
or whether to provide the cell content as seen by a spreadsheet user.}
}
\value{
A tibble representing the original spreadsheet format.
}
\description{
\code{simple_rectify} converts a set of cells  extracted by \code{\link[=read_ods_cells]{read_ods_cells()}}
into a traditional rectangular two-dimensional structure.
}
\details{
When set to \code{TRUE} (the default) the \code{base_values} argument then the
underlying cell values are used. When \code{FALSE} then cell content as seen by a
spreadsheet application user will be shown. See \code{\link[=read_ods_cells]{read_ods_cells()}} for further
details.

Column headers are of the format "X#", i.e. "X" and the column number. You
can use another function such as \code{\link[janitor:row_to_names]{janitor::row_to_names()}} to replace these
with values from a row within the dataset.
}
\examples{
example <- system.file("extdata", "basic_example.ods", package = "tidyods")
example_cells <- read_ods_cells(example, 1)
simple_rectify(example_cells, base_values = TRUE)

}
