% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyods-package.R
\docType{package}
\name{tidyods-package}
\alias{tidyods}
\alias{tidyods-package}
\title{tidyods: Read cells from ODS files}
\description{
Import cells from ODS files. Identify a cell's postion, value types and formulas, and provide methods to "rectify" cells back to a 2-dimensional data.frame.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattkerlogue/tidyods}
  \item \url{https://mattkerlogue.github.io/tidyods/}
  \item Report bugs at \url{https://github.com/mattkerlogue/tidyods/issues}
}

}
\author{
\strong{Maintainer}: Matt Kerlogue \email{matt.kerlogue@me.com}

}
\keyword{internal}
