% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_ods_cells}
\alias{read_ods_cells}
\title{Read cells from an ODS sheet}
\usage{
read_ods_cells(path, sheet = 1, quick = FALSE, whitespace = FALSE)
}
\arguments{
\item{path}{The ODS file}

\item{sheet}{The sheet within the ODS file}

\item{quick}{Whether to use the quick reading process}

\item{whitespace}{Whether to process multiple whitespaces}
}
\value{
A tibble with the following information:
\itemize{
\item \code{row}: the row number of the cell
\item \code{col}: the column number of the cell
\item \code{cell_type}: the type of cell, either "cell", "empty" or "merged"
\item \code{value_type}: the ODS value type (see details)
\item \code{cell_formula}: the ODS formula specification
\item \code{cell_content}: the value of the cell as represented to the user of a
spreadsheet application (i.e. with text formatting applied)
\item \code{base_value}: the underlying value of a cell (see details)
\item \verb{currency symbol}: the currency symbol for cells with a \code{value_type} of
currency
}

If using the \code{quick} argument you only \code{row}, \code{col} and \code{cell_content} are
returned.
}
\description{
Import cells from an OpenDocument Spreadsheet (ODS) file. In the resulting
dataaset each cell is its own row with columns providing information about
the cell's postion, value types and containing formulas.
}
\details{
Cells are assigned a \code{cell_type} of "cell" when they have content, "empty"
if they have no content, or "merged" if they are part of a set of merged cells.

The \code{value_type} of a cell is based on the \href{https://docs.oasis-open.org/office/OpenDocument/v1.3/os/part3-schema/OpenDocument-v1.3-os-part3-schema.html#__RefHeading__1417680_253892949}{"value-type" attribute}
defined by the OpenDocument Format specification. There are 7 value types:
\itemize{
\item \code{boolean}: equivalent to R's \code{\link[base:logical]{base::logical()}} data type
\item \code{currency}: a numeric value, optionally with a currency symbol
\item \code{date}: a date or date-time stored in ISO format
\item \code{float}: a numeric value, the ODS format does not distinguish between
different numeric types
\item \code{percentage}: a numeric value, formatted for display as a percentage
\item \code{string}: equivalent to R's \code{\link[base:character]{base::character()}}
\item \code{time}: a duration, stored in ISO format
}

For currency value types optionally a currency symbol (e.g. GBP, EUR, USD) can
be set, this is provided separate from the \code{base_value} as \code{currency_symbol}.

\code{read_ods_cells} provides two representations of a cell's value. The
\code{cell_content} column provides the cell's value as presented to the user of
a traditional spreadsheet application (such as Microsoft Excel, Google Sheets
or LibreOffice Calc), i.e. after applying cell formatting rules. The
\code{base_value} column provides the cell's underlying value, i.e. without
formatting.

Processing the ODS XML is a memory intensive process, you can achieve
significant speed enhancements by setting the \code{quick} argument to FALSE,
this will extract only the text content of the cell. Note this also ignores
repeated white space elements (i.e. \code{"A   B   C"} will be returned as
\code{"A B C"}).
}
\examples{
example <- system.file("extdata", "basic_example.ods", package = "tidyods")
example_cells <- read_ods_cells(example, 1)
dplyr::glimpse(example_cells)

}
