% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_ods_sheet}
\alias{read_ods_sheet}
\title{Read an ODS sheet to a rectangular dataset}
\usage{
read_ods_sheet(
  path,
  sheet,
  rectify = "simple",
  base_values = TRUE,
  quick = FALSE,
  whitespace = FALSE
)
}
\arguments{
\item{path}{The ODS file}

\item{sheet}{The sheet within the ODS file}

\item{rectify}{The method to convert cells to two-dimensions, can only be "simple"}

\item{base_values}{Whether to use the base_value of a cell (TRUE, the default)
or whether to provide the cell content as seen by a spreadsheet user.}

\item{quick}{Whether to use the quick reading process}

\item{whitespace}{Whether to process multiple whitespaces}
}
\value{
A tibble, presenting cells in a traditional two-dimension spreadsheet format.
}
\description{
A wrapper around \code{\link[=read_ods_cells]{read_ods_cells()}} and \code{\link[=simple_rectify]{simple_rectify()}} to extract
cells from a sheet and return a
}
\details{
At present, \code{rectify} can only be set to "simple", this calls the
\code{\link[=simple_rectify]{simple_rectify()}} method to convert the cells to a sheet.

When \code{base_values = TRUE} (the default) the underlying cell values are used
for non-string value types. When \code{FALSE} then cell content as seen by a
spreadsheet application user will be shown.

Setting \code{quick = TRUE} will ignore the setting of \code{base_values}, as the quick
process only returns cell content as seen by a spreadsheet users. When
\code{whitespace = FALSE} (the default), multiple whitespaces in cell content
are ignored. See \code{\link[=read_ods_cells]{read_ods_cells()}} for further details.
}
