% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectify.R
\name{simple_rectify}
\alias{simple_rectify}
\alias{smart_rectify}
\title{Convert ODS cells to a two-dimensional dataset}
\usage{
simple_rectify(ods_cells, skip = 0, col_headers = TRUE, base_values = TRUE)

smart_rectify(ods_cells)
}
\arguments{
\item{ods_cells}{A set of cells from \code{read_ods_cells()}.}

\item{skip}{The number of rows to skip before attempting to rectify the cells.}

\item{col_headers}{Whether to use the first row (after any skipping) as the
column header (\code{TRUE} is the default), alternatively a character vector of
column names can be provided.}

\item{base_values}{Whether to use the \code{base_value} of a cell (\code{TRUE}, the
default) or whether to provide the formatted \code{cell_content}, see
\code{\link[=read_ods_cells]{read_ods_cells()}} for further details.}
}
\value{
A tibble representing the original spreadsheet format.
}
\description{
There are two functions to "rectify" a set of ODS cells extracted by
\code{\link[=read_ods_cells]{read_ods_cells()}} back into a traditional two-dimensional spreadsheet
rectangle.
}
\details{
#' \code{simple_rectify()} will perform a basic reshaping of the dataset using either
the \code{base_value} or \code{cell_content} column of the supplied cells. It also
provides options to skip rows and whether to set column headers.

\code{smart_rectify()} performs a more complex reshaping of the dataset, by
guessing the location of column headers and using the \code{value_type} information
generated by \code{\link[=read_ods_cells]{read_ods_cells()}} to determine whether columns can be coerced
to a non-string data type (either numeric, logical, date or time).

When set to \code{TRUE} (the default) the \code{base_values} argument then the
underlying cell values are used. When \code{FALSE} then cell content as seen by a
spreadsheet application user will be shown. See \code{\link[=read_ods_cells]{read_ods_cells()}} for further
details.

If \code{col_headers = FALSE} or there is otherwise some other issue with
obtaining headers from the dataset then they will be are of the format
"X#", i.e. "X1", "X2", etc.
}
\examples{
example <- system.file("extdata", "basic_example.ods", package = "tidyods")
example_cells <- read_ods_cells(example, 1)
simple_rectify(example_cells, base_values = TRUE)
smart_rectify(example_cells)

}
